addressing equations_sbe
